assinatura e plano de assintura adiconado

+++ database/migrations/2023_02_08_133519_add_preco_to_servicos_table.php
@@ -14,7 +14,7 @@ class AddPrecoToServicosTable extends Migration
     public function up()
     {
         Schema::table('servicos', function (Blueprint $table) {
-            $table->integer('preco');
+            $table->double('preco');
         });
     }
